changing SmartHTTPInofRefs server component to work on strings and memory-based objects.
Previously tempfile.TemporaryFile() was used and was passed to subprocess.Popen as stdout outlet. That would involve extra file operations for temp files in cases when output is relatively small (hundreds of bytes). After the subprocess.PopebIO.communicateIO introduction, we can now reasonably easily manage the output in memory from start to finish.