Finalizing switch to CherryPy.WSGIServer as the default WSGI server.
The git_http_backend module is NOT tied to CherryPy's WSGI server. It can run against any WSGI 1.1-compatible (WSGI 1.0 if you don't work with git packs larger than 1Mb) server. The "default" means - when we run git_http_backend.py on the command-line - the simplest deployment scenario - it envokes a local copy of Cherrypy.WSGIServer and runs against it.